+++ fast_nav/lib/src/nested_navigator.dart
@@ -9,14 +9,14 @@ class NestedNavigator extends StatelessWidget {
   /// The name of the nested navigator
   final String name;
 
-  /// The home page of the navigator
-  final Widget home;
-
   /// [NavigatorObserver]s to be added to the nested navigator
   final Iterable<NavigatorObserver> observers;
 
   /// Called to generate a route for a given [RouteSettings].
   final RouteFactory? onGenerateRoute;
+
+  /// The home page of the navigator
+  final Widget? home;
 
   /// Create a nested navigator
   ///
@@ -25,9 +25,9 @@ class NestedNavigator extends StatelessWidget {
     Key? key,
     GlobalKey<NavigatorState>? navigatorKey,
     required this.name,
-    required this.home,
     this.observers = const [],
     this.onGenerateRoute,
+    this.home,
   })  : navigatorKey = navigatorKey ?? GlobalKey<NavigatorState>(),
         super(key: key);
 
@@ -36,9 +36,17 @@ class NestedNavigator extends StatelessWidget {
     return Navigator(
       key: FastNav.registerNavigator(name, key: navigatorKey),
       observers: [FastNavObserver(name), ...observers],
-      onGenerateRoute: (settings) =>
-          onGenerateRoute?.call(settings) ??
-          FastNav.generateAnonymousRoute(settings: settings, page: home),
+      onGenerateRoute: (settings) {
+        final route = onGenerateRoute?.call(settings);
+        if (route != null) {
+          return route;
+        } else if (home != null) {
+          return FastNav.generateAnonymousRoute(
+            settings: settings,
+            page: home!,
+          );
+        }
+      },
     );
   }
 }